examples/a2aadk/adk/requirements.txt
google-adk[a2a]
uvicorn
litellm